Extract PolicyListManager from PolicyList to its own file.
